fixing go modules
