--- serialreader/.gitignore
@@ -1,2 +1,0 @@
-.idea
-*.iml
