+++ src/datamodel/go.mod
@@ -1,0 +1,3 @@
+module com.github/cmyrsh/sm-reader/datamodel
+
+go 1.19
