+++ src/p1port/go.mod
@@ -1,0 +1,5 @@
+module com.github/cmyrsh/sm-reader/p1port
+
+replace com.github/cmyrsh/sm-reader/p1port => ../p1port
+
+go 1.19
